Add login dialog, update CAs, and wxs file
