+++ installers/msi-language/Login.wxs
@@ -1,0 +1,31 @@
+<?xml version="1.0" encoding="utf-8"?>
+<Wix xmlns="http://schemas.microsoft.com/wix/2006/wi">
+    <Fragment>
+        <UI Id="CustomUI_Login">
+            <Dialog Id="LoginDlg" Width="370" Height="270" Title="!(loc.LoginDlg_Title)">
+                <Control Id="BannerBitmap" Type="Bitmap" X="0" Y="0" Width="370" Height="44" TabSkip="no" Text="!(loc.PrivacyConsentDlgBannerBitmap)" />
+                <Control Id="BannerLine" Type="Line" X="0" Y="44" Width="370" Height="0" />
+                <Control Id="BottomLine" Type="Line" X="0" Y="234" Width="370" Height="0" />
+                <Control Id="Title" Type="Text" X="15" Y="6" Width="200" Height="15" Transparent="yes" NoPrefix="yes" Text="!(loc.LoginDlgTitle)" />
+                <Control Id="Next" Type="PushButton" X="236" Y="243" Width="56" Height="17" Default="yes" Text="!(loc.WixUINext)">
+                </Control>
+                <Control Id="Back" Type="PushButton" X="180" Y="243" Width="56" Height="17" Text="!(loc.WixUIBack)" />
+                <Control Id="Cancel" Type="PushButton" X="304" Y="243" Width="56" Height="17" Cancel="yes" Text="!(loc.WixUICancel)">
+                    <Publish Event="SpawnDialog" Value="CancelDlg">1</Publish>
+                </Control>
+                <Control Type="Edit" Id="UsernameEdit" Width="130" Height="15" X="68" Y="112" Property="AS_USERNAME" Indirect="no">
+                </Control>
+                <Control Type="Text" Id="PrivateProjectInfo" Width="263" Height="38" X="68" Y="56">
+                    <Text>You are attempting to install a private project, this requires that you are logged in. Please enter your username and password below.</Text>
+                </Control>
+                <Control Type="Text" Id="UsernameLabel" Width="47" Height="10" X="68" Y="98">
+                    <Text>Username</Text>
+                </Control>
+                <Control Type="Text" Id="PasswordLabel" Width="47" Height="11" X="68" Y="135">
+                    <Text>Password</Text>
+                </Control>
+                <Control Type="Edit" Id="PasswordEdit" Width="130" Height="15" X="68" Y="150" Property="AS_PASSWORD" Indirect="no" Password="yes" />
+            </Dialog>
+        </UI>
+    </Fragment>
+</Wix>
